Create smooth scroll container
To make sure that disabled eslint rules are for smooth scroll only
